Move Modal to component root
